+++ packages/bloc_test/lib/src/mock_bloc.dart
@@ -23,7 +23,62 @@ import 'package:mocktail/mocktail.dart';
 /// class MockCounterBloc extends MockBloc implements CounterBloc {}
 /// ```
 /// {@endtemplate}
-class MockBloc<E, S> extends Mock implements Bloc<E, S> {}
+class MockBloc<E, S> extends Mock implements Bloc<E, S> {
+  @override
+  dynamic noSuchMethod(Invocation invocation) {
+    try {
+      final dynamic result = super.noSuchMethod(invocation);
+      return result;
+    } on NoSuchMethodError {
+      final dynamic result = _noSuchMethod(invocation);
+      if (result == _unimplemented) rethrow;
+      return result;
+    }
+  }
+
+  dynamic _noSuchMethod(Invocation invocation) {
+    switch (invocation.memberName) {
+      case #listen:
+        return Stream<S>.empty().listen(
+          invocation.positionalArguments.first as void Function(S data),
+          onError: invocation.namedArguments[#onError] as Function?,
+          onDone: invocation.namedArguments[#onDone] as void Function()?,
+          cancelOnError: invocation.namedArguments[#cancelOnError] as bool?,
+        );
+      case #isBroadcast:
+        return true;
+      case #mapEventToState:
+        return Stream<S>.empty();
+      case #add:
+      case #onEvent:
+        return (E event) {}(invocation.positionalArguments.first as E);
+      case #addError:
+        return (Object error, [StackTrace? stackTrace]) {}(
+          invocation.positionalArguments.first as Object,
+          invocation.positionalArguments.length > 1
+              ? invocation.positionalArguments[1] as StackTrace?
+              : null,
+        );
+      case #onError:
+        return (Object error, StackTrace stackTrace) {}(
+          invocation.positionalArguments.first as Object,
+          invocation.positionalArguments[1] as StackTrace,
+        );
+      case #onTransition:
+        return (Transition<E, S> _) {}(
+          invocation.positionalArguments.first as Transition<E, S>,
+        );
+      case #emit:
+        return (S state) {}(invocation.positionalArguments.first as S);
+      case #close:
+        return Future<void>.value();
+      default:
+        return _unimplemented;
+    }
+  }
+}
+
+const _unimplemented = Object();
 
 /// {@template mock_cubit}
 /// Extend or mixin this class to mark the implementation as a [MockCubit].